+++ packages/theme/tasks/consumers.js
@@ -11,5 +11,7 @@ export default {
   'diagram-js': 'diagram-js/assets/diagram-js.css',
   'bpmn-js': 'bpmn-js/assets/bpmn-js.css',
   '@bpmn-io/properties-panel': 'properties-panel/src/assets/properties-panel.css',
-  'bpmn-js-element-templates': 'bpmn-js-element-templates/assets/element-templates.css'
+  'bpmn-js-element-templates': 'bpmn-js-element-templates/assets/element-templates.css',
+  '@bpmn-io/form-js-viewer': 'form-js/packages/form-js-viewer/assets/form-js-base.css',
+  '@bpmn-io/form-js-editor': 'form-js/packages/form-js-editor/assets/form-js-editor-base.css'
 };